Refactored normalization routines
