+++ tools/WorldEdit/src/App.tsx
@@ -68,10 +68,12 @@ import { ViewTabs } from "./components/ViewTabs";
 import { MapView } from "./components/MapView";
 import { ViewBody } from "./components/ViewBody";
 import {
+  classTitleCount,
   defaultSkillLevel,
   defaultSkillRating,
   groupSkillCount,
-  raceSkillCount
+  raceSkillCount,
+  raceStatKeys
 } from "./constants/globalDefaults";
 import {
   buildRoomRows,
@@ -176,6 +178,14 @@ import {
   tutorialColumns as tutorialColumnsDef
 } from "./crud/global/tutorialsCrud";
 import { createLoot, deleteLoot } from "./crud/global/lootCrud";
+import {
+  buildTitlePairs,
+  normalizeGroupSkills,
+  normalizeRaceClassMap,
+  normalizeRaceSkills,
+  normalizeRaceStats,
+  titlesToText
+} from "./utils/globalNormalizers";
 import type { VnumOption } from "./components/VnumPicker";
 import type { EventBinding } from "./data/eventTypes";
 import type {
@@ -338,9 +348,7 @@ const globalEntityOrder = [
 
 const primeStatOptions = ["str", "int", "wis", "dex", "con"] as const;
 const armorProfOptions = ["old_style", "cloth", "light", "medium", "heavy"] as const;
-const classTitleCount = 61;
 const classGuildCount = 2;
-const raceStatKeys = ["str", "int", "wis", "dex", "con"] as const;
 const lootEntryTypeOptions = ["item", "cp"] as const;
 const lootOpTypeOptions = [
   "use_group",
@@ -3350,98 +3358,6 @@ function buildExternalExits(
     }
     return a.direction.localeCompare(b.direction);
   });
-}
-
-function normalizeRaceStats(
-  value: RaceDefinition["stats"] | RaceDefinition["maxStats"]
-): Record<string, number> {
-  const stats: Record<string, number> = {
-    str: 0,
-    int: 0,
-    wis: 0,
-    dex: 0,
-    con: 0
-  };
-  if (Array.isArray(value)) {
-    raceStatKeys.forEach((key, index) => {
-      const parsed = Number(value[index]);
-      if (Number.isFinite(parsed)) {
-        stats[key] = parsed;
-      }
-    });
-    return stats;
-  }
-  if (value && typeof value === "object") {
-    raceStatKeys.forEach((key) => {
-      const record = value as Record<string, unknown>;
-      const parsed = Number(record[key]);
-      if (Number.isFinite(parsed)) {
-        stats[key] = parsed;
-      }
-    });
-  }
-  return stats;
-}
-
-function normalizeRaceClassMap(
-  value: Record<string, number> | number[] | undefined,
-  classNames: string[],
-  defaultValue: number
-): Record<string, number> {
-  const result: Record<string, number> = {};
-  if (!classNames.length) {
-    return result;
-  }
-  classNames.forEach((name, index) => {
-    let parsed: number | null = null;
-    if (Array.isArray(value)) {
-      parsed = Number(value[index]);
-    } else if (value && typeof value === "object") {
-      parsed = Number((value as Record<string, unknown>)[name]);
-    }
-    result[name] = Number.isFinite(parsed) ? Math.trunc(parsed) : defaultValue;
-  });
-  return result;
-}
-
-function normalizeRaceSkills(value: string[] | undefined): string[] {
-  return Array.from({ length: raceSkillCount }).map((_, index) => {
-    const entry = value?.[index];
-    return typeof entry === "string" ? entry : "";
-  });
-}
-
-function normalizeGroupSkills(value: string[] | undefined): string[] {
-  return Array.from({ length: groupSkillCount }).map((_, index) => {
-    const entry = value?.[index];
-    return typeof entry === "string" ? entry : "";
-  });
-}
-
-function titlesToText(titles: string[][] | undefined, column: 0 | 1): string {
-  const lines: string[] = [];
-  for (let i = 0; i < classTitleCount; i += 1) {
-    const value = titles?.[i]?.[column] ?? "";
-    lines.push(value);
-  }
-  return lines.join("\n");
-}
-
-function normalizeTitleLines(value: string | undefined): string[] {
-  if (!value) {
-    return [];
-  }
-  return value.split(/\r?\n/).map((line) => line.trim());
-}
-
-function buildTitlePairs(maleText: string, femaleText: string): string[][] {
-  const maleLines = normalizeTitleLines(maleText);
-  const femaleLines = normalizeTitleLines(femaleText);
-  const pairs: string[][] = [];
-  for (let i = 0; i < classTitleCount; i += 1) {
-    pairs.push([maleLines[i] ?? "", femaleLines[i] ?? ""]);
-  }
-  return pairs;
 }
 
 function cleanOptionalString(value: string | undefined): string | undefined {
