+++ tools/WorldEdit/src/constants/globalDefaults.ts
@@ -1,3 +1,5 @@
+export const classTitleCount = 61;
+export const raceStatKeys = ["str", "int", "wis", "dex", "con"] as const;
 export const raceSkillCount = 5;
 export const defaultSkillLevel = 53;
 export const defaultSkillRating = 0;
